requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.43.0
contourpy==1.3.2
cycler==0.12.1
Django==5.1.7
django-cloudinary-storage==0.3.0
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
idna==3.10
imageio==2.37.0
ipywidgets==8.1.2
Jinja2==3.1.6
jupyterlab_widgets==3.0.10
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.8.4
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
scikit-image==0.25.2
scipy==1.15.2
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
tifffile==2025.3.30
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.66.4
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.3.0
whitenoise==6.9.0
widgetsnbextension==4.0.10

gunicorn
psycopg2-binary